backend/requirements.txt
fastapi
uvicorn[standard]
langchain
pydantic
requests
beautifulsoup4
playwright
pandas
scikit-learn
plotly
pdfplumber
pypdf
pillow
python-dotenv
chromadb
websockets
httpx
beautifulsoup4
pytesseract
sentence_transformers
matplotlib
fpdf
reportlab
lxml
pdf2image
markdown
kaleido
